Renamed WindowClosedError.ts
