+++ packages/twitch-electron-auth-provider/src/WindowClosedError.ts
